cleanups and use `format_message` where possible
