+++ src/theme.rs
@@ -173,7 +173,9 @@ pub enum Container {
     Context,
     Highlight,
     SemiTransparent,
-    Custom { background: Option<Color> },
+    Custom {
+        background: Option<Color>,
+    },
 }
 
 impl container::StyleSheet for Theme {
@@ -246,7 +248,7 @@ impl container::StyleSheet for Theme {
             Container::Custom { background } => container::Appearance {
                 background: background.map(Background::Color),
                 ..Default::default()
-            }
+            },
         }
     }
 }
